feat: remove onClear prop on ChatWorkspace and replace with internal method
