+++ apps/src/aichat/views/ChatWorkspace.tsx
@@ -5,6 +5,7 @@ import React, {useCallback, useEffect, useMemo, useState} from 'react';
 import {useSelector} from 'react-redux';
 
 import {SystemPromptSettings} from '@cdo/apps/aichat/types';
+import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 import usePrevious from '@cdo/apps/util/usePrevious';
 
@@ -20,6 +21,7 @@ import {
   clearStagedFiles,
   fetchUserChatHistory,
   selectAllVisibleMessages,
+  sendAnalytics,
   setClientType,
   setNewChatSession,
 } from '../redux';
@@ -46,7 +48,6 @@ interface ChatWorkspaceProps {
   clientType: AiChatClientType;
   chatButtons?: ChatButtonAndKey[];
   hiddenContext?: string;
-  onClear: () => void;
 
   // Multimodal support
   multimodalEnabled?: boolean;
@@ -75,7 +76,6 @@ const ChatWorkspace: React.FunctionComponent<ChatWorkspaceProps> = ({
   clientType,
   chatButtons,
   hiddenContext,
-  onClear,
   multimodalEnabled = false,
   levelName,
   channelId,
@@ -260,6 +260,21 @@ const ChatWorkspace: React.FunctionComponent<ChatWorkspaceProps> = ({
     tabsContainerClassName: moduleStyles.tabsContainer,
     tabPanelsContainerClassName: moduleStyles.tabPanelsContainer,
   };
+
+  const onClear = useCallback(() => {
+    dispatch(clearChatMessages());
+    dispatch(
+      addChatEvent({
+        timestamp: Date.now(),
+        descriptionKey: 'CLEAR_CHAT',
+      })
+    );
+    dispatch(
+      sendAnalytics(EVENTS.CHAT_ACTION, {
+        action: 'Clear chat history',
+      })
+    );
+  }, [dispatch]);
 
   return (
     <div id="chat-workspace-area" className={moduleStyles.chatWorkspace}>
